+++ bold-pagos-en-linea/include/BoldGatewayBlocks.php
@@ -49,7 +49,7 @@ final class BoldGatewayBlocks extends AbstractPaymentMethodType {
 				'wp-html-entities',
 				'wp-i18n',
 			],
-			'3.1.0',
+			'3.1.1',
 			true
 		);
 		if ( function_exists( 'wp_set_script_translations' ) ) {