+++ bold-pagos-en-linea/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.1] - 2024-12-26
+### Added:
+- Add button payment Bold in list of editor blocks with Gutenberg.
+- Enable mode Embedded Checkout for payments buttons.
+
+### Changed:
+- Change design in selection of payment method in checkout.
+
 ## [3.1.0] - 2024-12-11
 ### Added:
 - Add support for payments in currency USD of store with conversion to COP.